feat: mve components and type to the package
