+++ apps/vardast-client/src/app/(client)/(profile)/profile/orders/[uuid]/products/components/OrderProductsTabs.tsx
@@ -10,6 +10,10 @@ import {
 } from "@vardast/graphql/generated"
 import { toast } from "@vardast/hook/use-toast"
 import graphqlRequestClientWithToken from "@vardast/query/queryClients/graphqlRequestClientWithToken"
+import {
+  CreateOrderLineSchema,
+  CreateOrderLineType
+} from "@vardast/type/OrderProductTabs"
 import zodI18nMap from "@vardast/util/zodErrorMap"
 import { ClientError } from "graphql-request"
 import { useForm, UseFormReturn } from "react-hook-form"
@@ -37,26 +41,6 @@ export enum OrderProductsTabsEnum {
   ORDER_MANUAL_TAB = "ORDER_MANUAL_TAB",
   UPLOAD_TAB_CONTENT = "UPLOAD_TAB_CONTENT",
   EXTRA_PRICE = "EXTRA_PRICE"
-}
-
-export const CreateOrderLineSchema = z.object({
-  attribuite: z.string().optional(),
-  brand: z.string(),
-  descriptions: z.string().optional(),
-  item_name: z.string(),
-  qty: z.string().optional(),
-  uom: z.string(),
-  type: z.string().optional()
-})
-
-export type CreateOrderLineType = TypeOf<typeof CreateOrderLineSchema> & {
-  type?: CreateLineInput["type"]
-}
-
-export interface OrderProductTabContentProps {
-  addProductLine: (productLine: CreateOrderLineType) => void
-  form: UseFormReturn<CreateOrderLineType>
-  uuid?: string
 }
 
 const OrderProductsTabs = ({ uuid }: OrderProductsTabsProps) => {
